WIP: gdiplus: Prevent gdi fonts from being free'd prematurely.
To prevent gdi from freeing (and then later re-loading from disk) fonts
that are still needed by some GpFont, add a HDC and HFONT to the GpFont
struct so fonts are never unused if some GpFont for it still exists.

When the GpFont is deleted, drop the HDC and HFONT so gdi can move the
font into the "unused cache" and eventually free it.

This feels like a terrible hack.